backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
mysql-connector-python==8.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pdfkit==1.0.0
openpyxl==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
requests==2.31.0